// C18-Charani-VLS-Capstone.zip/C18-Charani-VLS-FrontEnd-Project-Sln/src/app/model/data.model.ts
export interface Course {
  id?: string; // Make id optional for new courses (it will be auto-generated by the backend)
  name: string;
  description: string;
  trainer: string;
  price: number;
}

  export interface Author {
    id?: string; // Autogenerated ID, e.g., "A0001"
    name: string; // Author name
    course: string;
    description: string; // Brief description of the author
    rating: number;
  }

  export interface Learner {
    id?: string; // Autogenerated ID, e.g., "L0001"
    name: string; // Learner's name
    email: string;
    course: string;
    trainer: string;
    paymentmethod: string;
  }

  export interface Registration {
    id?: string; // Autogenerated ID, e.g., "R0001"
    name: string; // Learner's name
    email: string;
    phone: string;
    password: string; 
  }

  export interface Review {
    id?: string; // Autogenerated ID, e.g., "V0001"
    name: string;
    course: string; // Course being reviewed
    review: string; // Review text
    rating: number;
  }
  